Make "1" branch redundant, add quantity method
Identify and add quantity method, used flocking rules to get there.
Something to note from the book:

```
When the else branch is implemented first, :FIXME can always be used for the default. This not only saves you from having to figure out the right value, it also serves as a reminder to remove this temporary default later. If the non- else branch is implemented first, the default has to be set to something that actually meets the condition and so makes the true branch execute. Therefore, implementing the non- else branch first places a slightly greater burden on you. You have to use a specific, real value for the default, and then remember to remove the default once the transition is complete.
```

bookmark: 4.5 on page 87